refactor: 내부 클래스 이동
